% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AalenTimeDependentModel.R
\docType{class}
\name{AalenTimeDependentModel}
\alias{AalenTimeDependentModel}
\title{Models the relationship between a polygenic score and a phenotype calculated
through Aalen's additive hazards model}
\usage{
AalenTimeDependentModel(model, cumulative_effects, cumulative_variances)
}
\arguments{
\item{model}{The underlying model}

\item{cumulative_effects}{Numeric vector of cumulative effects}

\item{cumulative_variances}{Numeric vector of cumulative variances}
}
\description{
Models the relationship between a polygenic score and a phenotype calculated
through Aalen's additive hazards model
}
\section{Slots}{

\describe{
\item{\code{model}}{The Aalen model of the relationship}

\item{\code{effectFunction}}{Function which can be used to retrieve the effect at a
certain time point}
}}

\seealso{
\code{\link[=LoessTimeDependentModel]{LoessTimeDependentModel()}}, \code{\link[=GlmTimeDependentModel]{GlmTimeDependentModel()}},
\code{\link[timereg:aalen]{timereg::aalen()}}
}
