% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_MR.R
\name{calculate_mr_variance_trapz}
\alias{calculate_mr_variance_trapz}
\title{Calculate time-dependent effect variance of MR effects estimated by
integration using the trapezoidal rules}
\usage{
calculate_mr_variance_trapz(effect, variance)
}
\arguments{
\item{effect}{Numeric vector of total genetic effects on exposure for each
time point}

\item{variance}{Numeric vector of variance of genetic effect on outcome for
each time point.}
}
\value{
A numeric vector of variance at each time point
}
\description{
Calculate time-dependent effect variance of MR effects estimated by
integration using the trapezoidal rules
}
\keyword{internal}
