% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{derive_without_noise}
\alias{derive_without_noise}
\title{Derives cumulative effects where noise has been added to break ties}
\usage{
derive_without_noise(x, y)
}
\arguments{
\item{x}{The parameter of the function (usually time). The fractional part
of the values are assumed to be noise.}

\item{y}{The values of the function (usually cumulative effect)}
}
\value{
A numerically-derived function
}
\description{
Derives cumulative effects where noise has been added to break ties
}
\keyword{internal}
