% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmTimeDependentModel.R
\docType{methods}
\name{exponents}
\alias{exponents}
\alias{exponents,GlmTimeDependentModel-method}
\title{The exponents of age in the interaction terms. Same number and order as
\code{interactionEffects}}
\usage{
exponents(this)

\S4method{exponents}{GlmTimeDependentModel}(this)
}
\arguments{
\item{this}{The object to get the exponents from}
}
\description{
The exponents of age in the interaction terms. Same number and order as
\code{interactionEffects}
}
