% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmTimeDependentModel.R
\docType{methods}
\name{fixedEffect}
\alias{fixedEffect}
\alias{fixedEffect,GlmTimeDependentModel-method}
\title{Strength of the fixed effect}
\usage{
fixedEffect(this)

\S4method{fixedEffect}{GlmTimeDependentModel}(this)
}
\arguments{
\item{this}{The model to get the fixed effect strength from}
}
\description{
Strength of the fixed effect
}
