% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmTimeDependentModel.R
\docType{class}
\name{GlmTimeDependentModel-class}
\alias{GlmTimeDependentModel-class}
\alias{GlmTimeDependentModel}
\title{Models the relationship between age and time-dependent genetic effects using
a glm with interaction terms}
\description{
Models the relationship between age and time-dependent genetic effects using
a glm with interaction terms
}
\section{Slots}{

\describe{
\item{\code{model}}{The underlting model (\code{\link[stats:glm]{glm}})}

\item{\code{fixedEffect}}{The time-independent fixed genetic effect
(\code{numeric} length 1)}

\item{\code{interactionEffects}}{The coefficients of the interaction terms (\code{numeric})}

\item{\code{exponents}}{The exponent of age in the interaction terms (\code{numeric})}
}}

