% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmTimeDependentModel.R
\name{fixedEffect}
\alias{fixedEffect}
\alias{interactionEffects}
\alias{exponents}
\title{Strength of the fixed effect}
\usage{
fixedEffect(this)

interactionEffects(this)

exponents(this)
}
\description{
Strength of the fixed effect

Effect strength of the interaction terms. Same numer and order as \code{exponents}

The exponents of age in the interaction terms. Same number and order as
\code{interactionEffects}
}
\keyword{internal}
