% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_MR.R
\name{integrate_estimates}
\alias{integrate_estimates}
\title{Numerically integrate effect estimates by age}
\usage{
integrate_estimates(age, beta, method = c("trapezoidal", "midpoint"))
}
\arguments{
\item{age}{Numeric vector containing ages}

\item{beta}{Numeric vector containing effect estimates}

\item{method}{The integration method to use. Either \code{trapezoidal} (default)
or \code{midpoint}.}
}
\value{
A numeric matrix with columns age and cumulative integral
}
\description{
Numerically integrate effect estimates by age
}
\keyword{internal}
