% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_estimates.R
\name{integrate_trapezoid}
\alias{integrate_trapezoid}
\title{Integrate time-dependent effects using the trapezoidal rule and compute 95\%
CIs}
\usage{
integrate_trapezoid(age, wald_ratio, exposure_effect, outcome_variance)
}
\arguments{
\item{age}{The ages at which to integrate}

\item{wald_ratio}{The calculated Wald ratios}

\item{exposure_effect}{The genetic effect on the exposure at every age}

\item{outcome_variance}{The variance of the genetic effect on the outcome
at every age}
}
\value{
A data.frame with age, effect (Gamma), variance and 95\% CIs
}
\description{
Integrate time-dependent effects using the trapezoidal rule and compute 95\%
CIs
}
\keyword{internal}
