% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlmTimeDependentModel.R
\docType{methods}
\name{interactionEffects}
\alias{interactionEffects}
\alias{interactionEffects,GlmTimeDependentModel-method}
\title{Effect strengths of the interaction terms. Same numer and order as \code{exponents}}
\usage{
interactionEffects(this)

\S4method{interactionEffects}{GlmTimeDependentModel}(this)
}
\arguments{
\item{this}{The object to get the interaction effect strengths from}
}
\description{
Effect strengths of the interaction terms. Same numer and order as \code{exponents}
}
