% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last}
\alias{last}
\title{Get last element of a vector-like object}
\usage{
last(x)
}
\arguments{
\item{x}{Vector or list or anything that works with \code{\link[=tail]{tail()}}.}
}
\value{
The last element of \code{x}
}
\description{
Get last element of a vector-like object
}
\examples{
last(1:3)
}
\keyword{internal}
