% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessTimeDependentModel.R
\docType{class}
\name{LoessTimeDependentModel-class}
\alias{LoessTimeDependentModel-class}
\alias{LoessTimeDependentModel}
\title{Models the relationship between age and time-dependent genetic effects as
estimated through a LOESS model}
\description{
Models the relationship between age and time-dependent genetic effects as
estimated through a LOESS model
}
\section{Slots}{

\describe{
\item{\code{model}}{The LOESS model of the the PGS ~ age relationship}

\item{\code{stratifiedModels}}{Age-stratified \code{\link[stats:glm]{glm}} objects used to
estimate genetic effects.}
}}

\seealso{
\code{\link[=GlmTimeDependentModel]{GlmTimeDependentModel()}}, \code{\link[=time_dependent_loess]{time_dependent_loess()}}, \code{\link[stats:loess]{stats::loess()}}
}
