% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessTimeDependentModel.R
\name{stratifiedModels}
\alias{stratifiedModels}
\title{Named list of age-stratified models used to estimate effect strength for the
LOESS model}
\usage{
stratifiedModels(this)
}
\description{
Named list of age-stratified models used to estimate effect strength for the
LOESS model
}
\keyword{internal}
