% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentModel-class.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,TimeDependentModel-method}
\title{Return the underlying model of this instance}
\usage{
model(this)

\S4method{model}{TimeDependentModel}(this)
}
\arguments{
\item{this}{The object to get the model from.}
}
\description{
Return the underlying model of this instance
}
