% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentModel-class.R
\name{model}
\alias{model}
\title{Return the underlying model of this instance}
\usage{
model(this)
}
\arguments{
\item{this}{The object from which to extract the model from.}
}
\description{
Return the underlying model of this instance
}
