% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_all_in_one.R
\name{remove_outcome_cases}
\alias{remove_outcome_cases}
\title{Removes individuals with prior outcome events}
\usage{
remove_outcome_cases(
  data,
  outcome = "outcome",
  outcome_age = "outcome_age",
  exposure_age = "exposure_age"
)
}
\arguments{
\item{data}{A data frame containing the data set to be analyzed}

\item{outcome}{Column name of outcome status (i.e. had event or not)}

\item{outcome_age}{Column name of the age when the event occured}

\item{exposure_age}{Column name of the age when the exposure was measured}
}
\value{
A subset data frame where all observations where the outcome appeared
before the exposure were removed
}
\description{
Removes individuals with prior outcome events
}
\examples{
# Random example data
library(stats)
data <- data.frame(
  pgs = rnorm(50),
  outcome = sample(c(TRUE,FALSE), 50, replace=TRUE),
  outcome_age = rnorm(50, mean = 50, sd = 5) |> pmax(20),
  expoure_age = rnorm(50, mean = 30, sd = 2) |> pmax(15)
)

# Columns are already named the same as default values
remove_outcome_cases <- function(data)

# Case where one column has a different name
data$event <- data$outcome
data$outcome <- NULL
remove_outcome_cases <- function(data, outcome = "event")
}
