% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_all_in_one.R
\name{select_model}
\alias{select_model}
\title{Internal function for model selection}
\usage{
select_model(model_type, pgs, pheno, age, covariates, ...)
}
\arguments{
\item{model_type}{The model to run}

\item{...}{Arguments which are passed to the model}
}
\value{
The desired model
}
\description{
Internal function for model selection
}
\keyword{internal}
