% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoessTimeDependentModel.R
\docType{methods}
\name{stratifiedModels}
\alias{stratifiedModels}
\alias{stratifiedModels,LoessTimeDependentModel-method}
\title{Named list of age-stratified models used to estimate effect strength for the
LOESS model}
\usage{
stratifiedModels(this)

\S4method{stratifiedModels}{LoessTimeDependentModel}(this)
}
\arguments{
\item{this}{The instance to extract the models from}
}
\description{
Named list of age-stratified models used to estimate effect strength for the
LOESS model
}
