% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependent_aalen.R
\name{time_dependent_aalen}
\alias{time_dependent_aalen}
\title{Calculate time-dependent effect of polygenic scores on hazard}
\usage{
time_dependent_aalen(pgs, event, event_age, covariates)
}
\arguments{
\item{pgs}{Numeric vector of polygenic scores}

\item{event}{Event status, coded as specified by \code{\link[survival:Surv]{survival::Surv()}}}

\item{event_age}{Numeric vector of age at which event occured}

\item{covariates}{Data frame of covariates}
}
\value{
An instance of \code{\link[=AalenTimeDependentModel]{AalenTimeDependentModel()}}
}
\description{
Calculate time-dependent effect of polygenic scores on hazard
}
\examples{
# Start with random example data
library(stats)
pgs <- rnorm(50)
event <- sample(c(TRUE,FALSE), 50, replace=TRUE)
event_age <- rnorm(50, mean=40, sd = 5) |> pmax(20)
covariates <- data.frame(covar = rnorm(50))

model <- time_dependent_aalen(pgs, event, event_age, covariates)
}
\seealso{
\code{\link[=AalenTimeDependentModel]{AalenTimeDependentModel()}}
\code{\link[=time_dependent_glm]{time_dependent_glm()}},
\code{\link[=time_dependent_loess]{time_dependent_loess()}},
\code{\link[timereg:aalen]{timereg::aalen()}}
}
