% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependent_glm.R
\name{time_dependent_glm}
\alias{time_dependent_glm}
\title{Calculate time-dependent effects of a PGS on an exposure using a glm with
interaction terms}
\usage{
time_dependent_glm(pgs, pheno, age, covariates, exponents = 1)
}
\arguments{
\item{pgs}{Numeric vector of polygenic scores}

\item{pheno}{Numeric vector of phenotype measurements}

\item{age}{Numeric vector of age at assessment}

\item{covariates}{Data frame of covariates}

\item{exponents}{A numeric vector of exponents for the PGS-Age interaction term
(Default 1)}
}
\value{
An instance of \code{\link{GlmTimeDependentModel}} containing a
\code{\link[stats]{glm}} modeling the relationship and extracted coefficients
}
\description{
Calculate time-dependent effects of a PGS on an exposure using a glm with
interaction terms
}
\details{
The relationship is always modeled as
\code{exposure ~ pgs + age + covariates + interaction_terms} where
\code{interaction_terms} is an arbitrary number of PGS-Age interactions.
These interactions take the form \code{pgs:(age^exponent)} for each
exponent in \code{exponents}. Therefore, it is possible to include multiple
interaction terms to model, e.g., non-linear interactions. A value of
\code{NULL} causes no interaction term to be included.
}
\examples{
# Generate random example data
library(stats)
pgs <- rnorm(50)
pheno <- rnorm(50)
age <- rnorm(50, mean = 50, sd = 5) |> pmax(20)
covariates <- data.frame(covar = rnorm(50))

# Run model with linear pgs-age interaction
time_dependent_glm(pgs, pheno, age, covariates)
# Model with quadratic interaction
time_dependent_glm(pgs, pheno, age, covariates, exponents = 2)
# Model with linear and quartic interaction
time_dependent_glm(pgs, pheno, age, covariates, exponents = c(1,4))
# Model with no pgs-age interaction
time_dependent_glm(pgs, pheno, age, covariates, exponents = NULL)
}
