% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dependent_loess.R
\name{time_dependent_loess}
\alias{time_dependent_loess}
\title{Calculate time-dependent effects of a polygenic score on a phenotype using
a LOESS model}
\usage{
time_dependent_loess(
  pgs,
  pheno,
  age,
  covariates,
  age_range = NULL,
  age_step = 1
)
}
\arguments{
\item{pgs}{Numeric vector of polygenic scores}

\item{pheno}{Numeric vector of phenotype measurements}

\item{age}{Numeric vector of age at assessment}

\item{covariates}{Data frame of covariates}

\item{age_range}{Age range (min, max) for effect estimation or \code{NULL} for
automatic detection}

\item{age_step}{Number of years between each age stratum (default 1)}
}
\value{
An instance of \code{\link{LoessTimeDependentModel}} containing a
model of the relationship between age and genetic effect and all
age-stratified models generated in the calculation
}
\description{
Calculate time-dependent effects of a polygenic score on a phenotype using
a LOESS model
}
\details{
By default, \code{age_range} is inferred to be \code{range(data[, age])},
i.e. span from the minimum to the maximum age. The range is start- and
end-\emph{inclusive}. If the value does not represent a valid range, i.e.
element 2 is smaller than element 1, an error will be emitted. If the age
range only includes one age (i.e. element 1 = element 2), you will be warned
but the regression is still performed.
}
\examples{
# Generate random example data
library(stats)
pgs <- rnorm(1000)
pheno <- pgs * 3
age <- rep(40:70, length.out = 1000)
covariates <- data.frame(covar = rnorm(1000))

# Infer age range from data
time_dependent_loess(pgs, pheno, age, covariates)

# Use explicit age range from 50 to 65
time_dependent_loess(pgs, pheno, age, covariates, age_range = c(50,65))

# Only estimate effects for every other year
time_dependent_loess(pgs, pheno, age, covariates, age_step = 2)
}
\seealso{
\code{\link[=LoessTimeDependentModel]{LoessTimeDependentModel()}}, \code{\link[stats:loess]{stats::loess()}}, \code{\link[=time_dependent_glm]{time_dependent_glm()}}
}
