% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_MR.R
\name{time_dependent_MR}
\alias{time_dependent_MR}
\title{Perform time-dependent MR analysis}
\usage{
time_dependent_MR(
  age_seq,
  exposure_model,
  outcome_model,
  method = c("trapezoidal", "midpoint")
)
}
\arguments{
\item{age_seq}{Ages at which to estimate exposure-outcome effect}

\item{exposure_model}{Time-dependent model
(\code{\link[=GlmTimeDependentModel]{GlmTimeDependentModel}} or
\code{\link[=LoessTimeDependentModel]{LoessTimeDependentModel}}) of genetic effects
on exposure}

\item{outcome_model}{Time-dependent model
(\code{\link[=AalenTimeDependentModel]{AalenTimeDependentModel}}) of genetic effects
on outcome}

\item{method}{Integration method for calculating time-dependente
cumulative effects. \code{trapezoidal} (default) or \code{midpoint}.}
}
\value{
A data frame with the following columns:
\describe{
\item{age}{Time point of estimated effect}
\item{Gamma}{Cumulative effect}
\item{var}{Variance of cumulative effect}
\item{L95}{Lower end of 95\% CI}
\item{H95}{Higher end of 95\% CI}
}
}
\description{
Perform time-dependent MR analysis
}
\examples{
# Generate random example data
library(stats)
pgs <- rnorm(1000)
exposure <- pgs * 3
exposure_age <- rep(40:70, length.out = 1000)
covariates <- data.frame(covar = rnorm(1000))
outcome <- sample(c(TRUE,FALSE), 1000, replace=TRUE)
outcome_age <- rnorm(1000, mean=60, sd = 3) |> pmax(40)

# Estimate exposure and outcome effects
exposure_model <- time_dependent_loess(pgs, exposure, exposure_age, covariates)
outcome_model <- time_dependent_aalen(pgs, outcome, outcome_age, covariates)
time_dependent_MR(40:70, exposure_model, outcome_model)
}
\seealso{
\code{\link[=analyze_all_in_one]{analyze_all_in_one()}}, \code{\link[=time_dependent_aalen]{time_dependent_aalen()}},
\code{\link[=time_dependent_glm]{time_dependent_glm()}}, \code{\link[=time_dependent_loess]{time_dependent_loess()}},
\code{\link[=GlmTimeDependentModel]{GlmTimeDependentModel}},
\code{\link[=LoessTimeDependentModel]{LoessTimeDependentModel}},
\code{\link[=AalenTimeDependentModel]{AalenTimeDependentModel}}
}
