% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentModel-class.R
\docType{class}
\name{TimeDependentModel-class}
\alias{TimeDependentModel-class}
\alias{TimeDependentModel}
\title{Abstract class containing time-dependent effects for MR analysis}
\description{
Abstract class containing time-dependent effects for MR analysis
}
