% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeResolvedMR-package.R
\docType{package}
\name{TimeResolvedMR-package}
\alias{TimeResolvedMR}
\alias{TimeResolvedMR-package}
\title{TimeResolvedMR: Time-Resolved Mendelian Randomization Analysis}
\description{
Performs Mendelian Randomization analyses accounting for time-dependent effects of the exposure on the outcome.
}
\author{
\strong{Maintainer}: Daniel Schmitz \email{daniel.schmitz@igp.uu.se}

Authors:
\itemize{
  \item Torgny Karlsson \email{torgny.karlsson@igp.uu.se}
}

}
\keyword{internal}
