% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentModel-class.R,
%   R/AalenTimeDependentModel.R, R/GlmTimeDependentModel.R,
%   R/LoessTimeDependentModel.R
\docType{methods}
\name{totalEffect}
\alias{totalEffect}
\alias{totalEffect,TimeDependentModel-method}
\alias{totalEffect,AalenTimeDependentModel-method}
\alias{totalEffect,GlmTimeDependentModel-method}
\alias{totalEffect,LoessTimeDependentModel-method}
\title{Calculate the total genetic effect at a certain timepoint}
\usage{
totalEffect(this, age)

\S4method{totalEffect}{TimeDependentModel}(this, age)

\S4method{totalEffect}{AalenTimeDependentModel}(this, age)

\S4method{totalEffect}{GlmTimeDependentModel}(this, age)

\S4method{totalEffect}{LoessTimeDependentModel}(this, age)
}
\arguments{
\item{this}{The model to use for calculation}

\item{age}{The ages (or time points) for which to calculate the genetic effect}
}
\description{
Calculate the total genetic effect at a certain timepoint
}
