% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDependentModel-class.R
\name{totalEffect}
\alias{totalEffect}
\title{Calculate the total genetic effect at a certain timepoint}
\usage{
totalEffect(this, age)
}
\arguments{
\item{this}{The model to use for calculation}

\item{age}{The ages (or time points) for which to calculate the genetic effect}
}
\description{
Calculate the total genetic effect at a certain timepoint
}
