% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AalenTimeDependentModel.R
\docType{methods}
\name{variance}
\alias{variance}
\alias{variance,AalenTimeDependentModel-method}
\title{Calculate hazard variance at a certain age}
\usage{
variance(this, age)

\S4method{variance}{AalenTimeDependentModel}(this, age)
}
\arguments{
\item{this}{Model to use for calcularion}

\item{age}{Numeric vector of time points to retrieve the variances for.}
}
\description{
Calculate hazard variance at a certain age
}
